apps/rag_lmbd_query/requirements.txt
boto3
python-dotenv
psycopg2-binary
numpy